apps/backend/requirements.txt
Django==4.2.21
django-constance==4.3.2
django-cors-headers==4.7.0
git+https://github.com/derneuere/django-chunked-upload@master#egg=django-chunked-upload
django-cryptography==1.1
django-extensions==3.2.3
django-filter==24.3
django-picklefield==3.3
django-bulk-update
django-silk==5.3.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
face-recognition==1.3.0
faiss-cpu==1.10.0
Flask==3.1.1
Flask-Cors==6.0.0
Flask-RESTful==0.3.10
geopy==2.4.1
gunicorn==23.0.0
hdbscan==0.8.40
matplotlib==3.10.1
networkx==3.4.2
nltk==3.9.1
markupsafe==3.0.2
Pillow==10.4.0
psycopg==3.2.6
https://github.com/owncloud/pyocclient/archive/master.zip
pytz==2024.2
tzdata==2024.2
PyExifTool==0.4.9
pyvips==2.2.3
scikit-learn<1.6.2
seaborn==0.13.2
sentence_transformers==2.7.0
timezonefinder==6.5.9
tqdm==4.67.1
gevent==24.11.1
python-magic==0.4.27
Wand==0.6.13
django-q2==1.7.6
safetensors==0.5.3
py-cpuinfo==9.0.0
psutil==6.1.1

# Dependencies for blip
timm==1.0.15
# Dependencies for Moondream chat handler (llama-cpp-python multimodal)
transformers==4.51.3
# Dependencies for mistral quantized and multimodal models like Moondream
llama-cpp-python==0.3.9
